+++ source/extensions/omni.isaac.ros2_bridge-humble/nodes/OgnROS2PublishJointState.cpp
@@ -46,7 +46,18 @@ public:
         const GraphContextObj& context = db.abi_context();
 
         auto& state = db.internalState<OgnROS2PublishJointState>();
-        const char* primPath = db.inputs.targetPrim.path();
+
+        const auto& prim = db.inputs.targetPrim();
+        const char* primPath;
+        if (prim.size() > 0)
+        {
+            primPath = omni::fabric::toSdfPath(prim[0]).GetText();
+        }
+        else
+        {
+            db.logError("no prim path found");
+            return false;
+        }
 
         // spin once calls reset automatically if it was not successful
         const auto& nodeObj = db.abi_node();