+++ source/extensions/omni.isaac.examples_nodes/docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
-
+## [0.1.2] - 2023-08-10
+### Changed
+- Changed robot prim type from bundle to target
 ## [0.1.1] - 2023-06-12
 ### Changed
 - Update to kit 105.1